Calibrated Fisher Z.
