+++ tetrad-lib/src/main/java/edu/cmu/tetrad/algcomparison/independence/SemBicTest.java
@@ -19,7 +19,7 @@ import java.util.*;
         command = "sem-bic-test",
         dataType = {DataType.Continuous, DataType.Covariance}
 )
-@Experimental
+//@Experimental
 public class SemBicTest implements IndependenceWrapper {
 
     static final long serialVersionUID = 23L;
